+++ guard/lib/guard/api/github.ex
@@ -40,14 +40,14 @@ defmodule Guard.Api.Github do
 
     case Cachex.get(:token_cache, cache_key) do
       {:ok, {token, expires_at}} when not is_nil(token) and token != "" ->
-        if valid_token?(expires_at) do
+        if OAuth.valid_token?(expires_at) do
           {:ok, {token, expires_at}}
         else
           handle_fetch_and_cache_token(repo_host_account)
         end
 
       _ ->
-        handle_fetch_and_cache_token(repo_host_account)
+        handle_static_token_or_fetch(repo_host_account)
     end
   end
 
@@ -76,7 +76,7 @@ defmodule Guard.Api.Github do
 
     case Tesla.post(client, @oauth_path, nil, query: query_params) do
       {:ok, %Tesla.Env{status: status, body: body}} when status in 200..299 ->
-        handle_ok_token_response(repo_host_account, body)
+        OAuth.handle_ok_token_response(repo_host_account, body)
 
       {:ok, %Tesla.Env{status: status}} when status in 400..499 ->
         Logger.warning("Failed to refresh github token, account might be revoked")
@@ -91,49 +91,11 @@ defmodule Guard.Api.Github do
     end
   end
 
-  defp handle_ok_token_response(repo_host_account, body) do
-    body =
-      if is_binary(body) do
-        Jason.decode!(body)
-      else
-        body
-      end
-
-    token = body["access_token"]
-    expires_in = body["expires_in"]
-    refresh_token = body["refresh_token"]
-
-    expires_at = calc_expires_at(expires_in)
-
-    if valid_token?(expires_at) do
-      Cachex.put(:token_cache, token_cache_key(repo_host_account.id), {token, expires_at})
-      update_refresh_token(repo_host_account, refresh_token)
-    end
-
-    {:ok, {token, expires_at}}
-  end
-
   defp build_token_client do
     Tesla.client([
       {Tesla.Middleware.BaseUrl, @oauth_base_url},
       Tesla.Middleware.JSON
     ])
-  end
-
-  # Case where the token never expires
-  defp valid_token?(nil), do: true
-
-  defp valid_token?(expires_at) do
-    current_time = DateTime.utc_now() |> DateTime.to_unix()
-    # 5 minutes before expiration
-    expires_at - 300 > current_time
-  end
-
-  defp calc_expires_at(nil), do: nil
-
-  defp calc_expires_at(expires_in) do
-    current_time = DateTime.utc_now() |> DateTime.to_unix()
-    current_time + expires_in
   end
 
   def validate_token(""), do: false
